backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3

# High Priority Upgrades
python-dotenv==1.0.1
python-json-logger==3.2.1
marshmallow-enum==1.5.1

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0
